+++ sei-ibc-go/modules/core/types/query.go
@@ -9,6 +9,8 @@ import (
 	connectiontypes "github.com/cosmos/ibc-go/modules/core/03-connection/types"
 	channel "github.com/cosmos/ibc-go/modules/core/04-channel"
 	channeltypes "github.com/cosmos/ibc-go/modules/core/04-channel/types"
+	port "github.com/cosmos/ibc-go/modules/core/05-port"
+	porttypes "github.com/cosmos/ibc-go/modules/core/05-port/types"
 )
 
 // QueryServer defines the IBC interfaces that the gRPC query server must implement
@@ -16,6 +18,7 @@ type QueryServer interface {
 	clienttypes.QueryServer
 	connectiontypes.QueryServer
 	channeltypes.QueryServer
+	porttypes.QueryServer
 }
 
 // RegisterQueryService registers each individual IBC submodule query service
@@ -23,4 +26,5 @@ func RegisterQueryService(server grpc.Server, queryService QueryServer) {
 	client.RegisterQueryService(server, queryService)
 	connection.RegisterQueryService(server, queryService)
 	channel.RegisterQueryService(server, queryService)
+	port.RegisterQueryService(server, queryService)
 }